MoppelHelper.ShowHideColumns storage;  Activity for friends compare;
Guild is now a "General" key in the i18n files.

In the compare friends thread the activity indicator is added.

Some of the columns (as of now: Guild, Era and Points) can be hidden; settings are saved to the local storage.

Variable name changes (Start/End instead of Min/Max) in the part-calc module.